Creating option to do not create the container
Creating the option to do not create the container for the output files
. Atempting to create the container was consuming one imense amount of
time when you are generate tasks for a job with a large numebr of files.
for a job with 2700 files I was able to reduce the task generation time
by including this option from 27 minutes to 2 minutes.